refactor: Remove unused and redundant code
