+++ frappe/core/doctype/prepared_report/prepared_report.py
@@ -11,8 +11,7 @@ from frappe.desk.query_report import generate_report_result
 from frappe.model.document import Document
 from frappe.utils import gzip_compress, gzip_decompress
 from frappe.utils.background_jobs import enqueue
-from frappe.core.doctype.file.file import remove_all
-
+from frappe.utils.file_manager import remove_all
 
 class PreparedReport(Document):
 	def before_insert(self):
