+++ frappe/website/doctype/personal_data_download_request/personal_data_download_request.py
@@ -28,7 +28,7 @@ class PersonalDataDownloadRequest(Document):
 		})
 		f.save(ignore_permissions=True)
 
-		file_link = frappe.utils.get_url("/api/method/frappe.core.doctype.file.file.download_file") +\
+		file_link = frappe.utils.get_url("/api/method/frappe.utils.file_manager.download_file") +\
 			"?" + get_signed_params({"file_url": f.file_url})
 		host_name = frappe.local.site
 		frappe.sendmail(
